Refactor of styles complete, need lots of css love
